Authorization don

+++ backend/routes/posts.js
@@ -38,9 +38,12 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
+      creator: req.userData.userId,
       imagePath: url + '/images/' + req.file.filename
     });
+    
     post.save().then(createdPost => {
+    
       res.status(201).json({
         message: 'Post added successfully',
         post: {
@@ -101,19 +104,29 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: 'Update successful!' });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      
+      if (result.nModified > 0) {
+        res.status(200).json({ message: 'Update successful!' });
+      } else {
+        res.status(401).json({ message: 'Not Authorized!' });
+      }
+
     });
   }
 );
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Post deleted!' });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Post deleted!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized!' });
+    }
   });
 });
